requirements.txt
Flask==2.2.3
pandas==1.5.3
python-dotenv==1.0.0
shroomdk==1.0.2
